Finish read csv service

+++ importFiles/src/main/java/com/datacloud/importFiles/dto/ProductDTO.java
@@ -8,9 +8,9 @@ import lombok.Setter;
 public class ProductDTO {
     private String name;
     private double price;
-    private int quantity;
+    private double quantity;
 
-    public ProductDTO(String name, double price, int quantity) {
+    public ProductDTO(String name, double price, double quantity) {
         this.name = name;
         this.price = price;
         this.quantity = quantity;
